+++ tests/run.php
@@ -121,6 +121,7 @@ testCodesCommandDoesNotRequireConfiguredSources($project, $binary);
 testStdinInputDoesNotRequireConfiguredSources($project, $root);
 testStagedInputDoesNotRequireConfiguredSources($project, $root);
 testPhpStanLevelAnalyzeRunsEndToEnd($project, $binary);
+testAnalyzeReportsMissingTypeHints($project, $binary);
 testPhpStanPragmaOverlayGeneration($project, $root);
 testLaravelPaginatorReturnDocblockOverlayGeneration($root);
 testLaravelDateHelperOverlayGeneration($project, $root);